Adding prettier configuration file for formatting

+++ src/contracts/Exchange.sol
@@ -4,76 +4,76 @@ import "./Token.sol";
 import "openzeppelin-solidity/contracts/math/SafeMath.sol";
 
 contract Exchange {
-    using SafeMath for uint;
+	using SafeMath for uint256;
 
-    // Variables
-    address public feeAccount; // the account that receives exchange fees
-    uint256 public feePercent; // the fee percentage
-    address constant ETHER = address(0); // store Ether in tokens mapping with blank address
-    mapping(address => mapping(address => uint256)) public tokens; // token address to user balance
+	// Variables
+	address public feeAccount; // the account that receives exchange fees
+	uint256 public feePercent; // the fee percentage
+	address constant ETHER = address(0); // store Ether in tokens mapping with blank address
+	mapping(address => mapping(address => uint256)) public tokens; // token address to user balance
 
-    // Events
-    event Deposit(
-        address _token,
-        address _user,
-        uint256 _amount,
-        uint256 _balance
-    );
+	// Events
+	event Deposit(
+		address _token,
+		address _user,
+		uint256 _amount,
+		uint256 _balance
+	);
 
-    event Withdraw(
-        address _token,
-        address _user,
-        uint256 _amount,
-        uint256 _balance
-    );
+	event Withdraw(
+		address _token,
+		address _user,
+		uint256 _amount,
+		uint256 _balance
+	);
 
-    constructor(address _feeAccount, uint256 _feePercent) public {
-        feeAccount = _feeAccount;
-        feePercent = _feePercent;
-    }
+	constructor(address _feeAccount, uint256 _feePercent) public {
+		feeAccount = _feeAccount;
+		feePercent = _feePercent;
+	}
 
-    // Fallback: reverts if Ether is sent to this smart contract by mistake
-    function() external {
-        revert();
-    }
+	// Fallback: reverts if Ether is sent to this smart contract by mistake
+	function() external {
+		revert();
+	}
 
-    function depositEther() public payable {
-        tokens[ETHER][msg.sender] = tokens[ETHER][msg.sender].add(msg.value);
-        emit Deposit(ETHER, msg.sender, msg.value, tokens[ETHER][msg.sender]);
-    }
+	function depositEther() public payable {
+		tokens[ETHER][msg.sender] = tokens[ETHER][msg.sender].add(msg.value);
+		emit Deposit(ETHER, msg.sender, msg.value, tokens[ETHER][msg.sender]);
+	}
 
-    function withdrawEther(uint256 _amount) public {
-        require(tokens[ETHER][msg.sender] >= _amount);
-        tokens[ETHER][msg.sender] = tokens[ETHER][msg.sender].sub(_amount);
+	function withdrawEther(uint256 _amount) public {
+		require(tokens[ETHER][msg.sender] >= _amount);
+		tokens[ETHER][msg.sender] = tokens[ETHER][msg.sender].sub(_amount);
 
-        // in newer solidity its recommended to use
-        // (bool success, ) = msg.sender.call{value: _amount}("");
-        // require(success, "Transfer failed.");
-        msg.sender.transfer(_amount);
+		// in newer solidity its recommended to use
+		// (bool success, ) = msg.sender.call{value: _amount}("");
+		// require(success, "Transfer failed.");
+		msg.sender.transfer(_amount);
 
-        emit Withdraw(ETHER, msg.sender, _amount, tokens[ETHER][msg.sender]);
-    }
+		emit Withdraw(ETHER, msg.sender, _amount, tokens[ETHER][msg.sender]);
+	}
 
-    function depositToken(address _token, uint256 _amount) public {
-        require(_token != ETHER);
-        require(Token(_token).transferFrom(msg.sender, address(this), _amount));
-        tokens[_token][msg.sender] = tokens[_token][msg.sender].add(_amount);
-        emit Deposit(_token, msg.sender, _amount, tokens[_token][msg.sender]);
-    }
+	function depositToken(address _token, uint256 _amount) public {
+		require(_token != ETHER);
+		require(Token(_token).transferFrom(msg.sender, address(this), _amount));
+		tokens[_token][msg.sender] = tokens[_token][msg.sender].add(_amount);
+		emit Deposit(_token, msg.sender, _amount, tokens[_token][msg.sender]);
+	}
 
-    function withdrawToken(address _token, uint256 _amount) public {
-        require(_token != ETHER);
-        require(tokens[_token][msg.sender] >= _amount);
-        tokens[_token][msg.sender] = tokens[_token][msg.sender].sub(_amount);
-        require(Token(_token).transfer(msg.sender, _amount));
-        emit Withdraw(_token, msg.sender, _amount, tokens[_token][msg.sender]);
-    }
+	function withdrawToken(address _token, uint256 _amount) public {
+		require(_token != ETHER);
+		require(tokens[_token][msg.sender] >= _amount);
+		tokens[_token][msg.sender] = tokens[_token][msg.sender].sub(_amount);
+		require(Token(_token).transfer(msg.sender, _amount));
+		emit Withdraw(_token, msg.sender, _amount, tokens[_token][msg.sender]);
+	}
 
-    function balanceOf(address _token, address _user)
-        public
-        view
-        returns (uint256 _balance)
-    {
-        return tokens[_token][_user];
-    }
+	function balanceOf(address _token, address _user)
+		public
+		view
+		returns (uint256 _balance)
+	{
+		return tokens[_token][_user];
+	}
 }
